fix error

+++ OpenRA.MapReader/NoiseGenerator.cs
@@ -78,7 +78,7 @@ namespace OpenRA.MapReader
             foreach (float wavelength in wavelengths)
             {
                 float amps = AmpFunc(wavelength);
-                int subSize = ((size / wavelength) | 0) + 2;
+                int subSize = ((int)((size / wavelength)) | 0) + 2;
                 float[] subNoise = GeneratePerlinNoise2D(subSize);
 
                 // Offsets should align to grid
